notification icon
icon only shows when app is collecting data and is removed when
collection of data stops.

fixed issue where icon was duplicated in notification area.

must be polished: clicking PPN (PedalPortland notification) in
notification drawer appears it may reset application but further
understanding of Android and our application is required.

We need to determine 1) what activity we want the application to do
when the PPN is selected 2) do we want to prevent the user from
removing the PPN until data is no longer being collected 3) which icon
to be used (if GPS notification is not desired)

This concludes issue #3 "Show status icon in status bar" for now.
To be revisited after further understanding and clarification evolves